Shader name refactor
- Error now calculated as predicted minus actual instead of actual minus predicted
-Renamed shaders
