+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust_perceptron"
-version = "0.3.5"
+version = "0.3.6"
 authors = ["gregoryquick <gregory@quicks.me>"]
 edition = "2018"
