+++ src/pipelines/applyweights.rs
@@ -1,11 +1,12 @@
 pub struct Pipeline {
     pub uniform_buffer: wgpu::Buffer,
-    pub input_buffer_a: wgpu::Buffer,
-    pub input_buffer_b: wgpu::Buffer,
+    pub weight_buffer: wgpu::Buffer,
+    pub input_buffer: wgpu::Buffer,
     pub output_buffer: wgpu::Buffer,
-    pub bind_group_0: wgpu::BindGroup,
-    pub compute_pipeline: wgpu::ComputePipeline,
+    bind_group_0: wgpu::BindGroup,
+    compute_pipeline: wgpu::ComputePipeline,
 }
+
 impl Pipeline {
     pub fn new<T: bytemuck::Pod>(anchor: &super::PipelineAnchor,
                                  buffers: (Option<wgpu::Buffer>,
@@ -13,11 +14,12 @@ impl Pipeline {
                                            Option<wgpu::Buffer>)
                                  , batch_size: usize,) -> Self {
         let type_size = std::mem::size_of::<T>();
+        let input_size = anchor.input_size;
         let output_size = anchor.output_size;
         let device = &anchor.device;
         use wgpu::util::{BufferInitDescriptor, DeviceExt};
         //Create buffers
-        let uniform_data = [output_size as u32, batch_size as u32];
+        let uniform_data = [input_size as u32, output_size as u32, batch_size as u32];
         let uniform_buffer = device.create_buffer_init(
             &BufferInitDescriptor {
                 label: Some("Uniform Buffer"),
@@ -27,30 +29,30 @@ impl Pipeline {
         );
         //0-0
 
-        let input_buffer_a = buffers.0.unwrap_or(
+        let weight_buffer = buffers.0.unwrap_or(
             device.create_buffer(
                 &wgpu::BufferDescriptor {
-                    label: Some("Input Buffer A"),
-                    size: (type_size * output_size * batch_size) as wgpu::BufferAddress,
+                    label: Some("Network Weights"),
+                    size: (type_size * input_size * output_size) as wgpu::BufferAddress,
                     usage: wgpu::BufferUsage::STORAGE | wgpu::BufferUsage::COPY_DST,
                     mapped_at_creation: false,
                 }
             )
         );
         //0-1
-        
-        let input_buffer_b = buffers.1.unwrap_or(
+
+        let input_buffer = buffers.1.unwrap_or(
             device.create_buffer(
                 &wgpu::BufferDescriptor {
-                    label: Some("Input Buffer B"),
-                    size: (type_size * output_size * batch_size) as wgpu::BufferAddress,
+                    label: Some("Input Buffer"),
+                    size: (type_size * input_size * batch_size) as wgpu::BufferAddress,
                     usage: wgpu::BufferUsage::STORAGE | wgpu::BufferUsage::COPY_DST,
                     mapped_at_creation: false,
                 }
             )
         );
         //0-2
-
+        
         let output_buffer = buffers.2.unwrap_or(
             device.create_buffer(
                 &wgpu::BufferDescriptor {
@@ -66,7 +68,7 @@ impl Pipeline {
         //Create bind group(s)
         let bind_group_layout_0 = device.create_bind_group_layout(
             &wgpu::BindGroupLayoutDescriptor {
-                label: Some("Leaky relu bind group layout 0"),
+                label: Some("Weight aplication bind group layout 0"),
                 entries: &[wgpu::BindGroupLayoutEntry {
                     binding: 0,
                     visibility: wgpu::ShaderStage::COMPUTE,
@@ -117,7 +119,7 @@ impl Pipeline {
         );
         let bind_group_0 = device.create_bind_group(
             &wgpu::BindGroupDescriptor {
-                label:  Some("Leaky relu bind group 0"),
+                label:  Some("Weight aplication bind group 0"),
                 layout: &bind_group_layout_0,
                 entries: &[wgpu::BindGroupEntry {
                     binding: 0,
@@ -125,13 +127,12 @@ impl Pipeline {
                 },
                 wgpu::BindGroupEntry {
                     binding: 1,
-                    resource: input_buffer_a.as_entire_binding(),
+                    resource: weight_buffer.as_entire_binding(),
                 },
-                                wgpu::BindGroupEntry {
+                wgpu::BindGroupEntry {
                     binding: 2,
-                    resource: input_buffer_b.as_entire_binding(),
+                    resource: input_buffer.as_entire_binding(),
                 },
-
                 wgpu::BindGroupEntry {
                     binding: 3,
                     resource: output_buffer.as_entire_binding(),
@@ -140,9 +141,9 @@ impl Pipeline {
         );
 
         //Create compute pipeline
-        let cs_src = include_str!("shaders/subtract.comp");
+        let cs_src = include_str!("shaders/applyweights.comp");
         let mut compiler = shaderc::Compiler::new().unwrap();
-        let cs_spirv = compiler.compile_into_spirv(cs_src, shaderc::ShaderKind::Compute, "subtract.comp", "main", None).unwrap();
+        let cs_spirv = compiler.compile_into_spirv(cs_src, shaderc::ShaderKind::Compute, "applyweights.comp", "main", None).unwrap();
         let cs_module = device.create_shader_module(
             &wgpu::ShaderModuleDescriptor {
                 label: None,
@@ -161,7 +162,7 @@ impl Pipeline {
 
         let compute_pipeline = device.create_compute_pipeline(
             &wgpu::ComputePipelineDescriptor {
-                label: Some("Leaky relu pipeline"),
+                label: Some("Weight aplication pipeline"),
                 layout: Some(&pipeline_layout),
                 module: &cs_module,
                 entry_point: "main",
@@ -172,14 +173,14 @@ impl Pipeline {
         //Return
         Pipeline {
             uniform_buffer,
-            input_buffer_a,
-            input_buffer_b,
+            weight_buffer,
+            input_buffer,
             output_buffer,
             bind_group_0,
             compute_pipeline,
         }
     }
-    
+
     pub fn run(&self, anchor: &super::PipelineAnchor, encoder: &mut wgpu::CommandEncoder, batch_size: usize) {
         //Create compute pass
         let mut compute_pass = encoder.begin_compute_pass(
